Add support for custom type translations and optional namespace
